+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.3 - 2026-07-16
+
+- Highlight IMDb ratings of 7.0 or higher in bold red in filtered RSS item descriptions.
+- Add `HIGHLIGHT_RATING` to configure the bold red IMDb rating threshold.
+
 ## 0.2 - 2026-06-18
 
 - Add IMDb search links to filtered RSS item descriptions.